refactor(services): add common errors
