+++ django/extractor/errors.py
@@ -36,11 +36,3 @@ class EmptyResult(Exception):
     """
 
     pass
-
-
-class BatchCancelled(Exception):
-    """
-    Error used when a batch has been cancelled.
-    """
-
-    pass
