+++ django/transformer/errors.py
@@ -1,10 +1,2 @@
 class OperationOutcome(Exception):
     pass
-
-
-class BatchCancelled(Exception):
-    """
-    Error used when a batch has been cancelled.
-    """
-
-    pass
